Changes for post/fetch requests to work
Made changes to detection.py for the variable names for what is being posted to the server.
Tried making changes to the app.js file to accept a different variable name but this is the only change that worked.
Note: Noticed that on the website the available parking spots only changes when a new lot is selected, possibly something to change/fix after the demo